correction_tool_backend/requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.6
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
drf-yasg
transformers
torch
django-cors-headers
django-allauth
cryptography
dj-rest-auth
pyspellchecker==0.7.2
